fix: move bip39 to its own module + write doc and comment
